Fix BtIO race condition in detecting a closed file descriptor
With the BtIO API the way to abort some operation is by calling
g_io_channel_shutdown. This will cause the corresponding file descriptor to be
closed and G_IO_NVAL to be reported to the BtIO internal watch callback. When
BtIO gets G_IO_NVAL it knows not to call back to the application since the
operation is considered aborted.

There's unfortunately a race condition associated with this. The G_IO_NVAL will
get reported only in the next main loop iteration that follows the one where
the file descriptor was closed. So, if there was input or an error for the file
descriptor in the same iteration where it was closed the BtIO internal watch
will get called with something other than G_IO_NVAL. In other words calling
g_io_channel_shutdown doesn't provide a 100% guarantee that the application
callback will not get called.

This patch fixes the problem by doing a secondary check for POLLNVAL by calling
poll() with a zero timeout in the BtIO internal watch functions.