Fix agent not being released after transfers complete
This requession was introduced by d57bffe46b71e17a640c11b389dd6da95f933729
that add another reference to the session for the agent.

To fix this a rework on refcount was done so that transfer now hold
references to the session and once done they release the references one
by one.