ebook backend: separate concatenated vcards with \r\n.
The END:VCARD is not ended with \r\n for single vcards.
In a phonebook pull request vcards are concatenated as one
big blob. Carkits seem to choke on the fact that an END:VCARD
is directly followed by BEGIN:VCARD without \r\n separation.
(as observed by many n900 users).